requirements.txt
args==0.1.0
asgiref==3.5.2
click==8.0.4
cmake==3.22.5
colorama==0.4.4
django-cors-headers==3.12.0
Django==4.0.4
djangorestframework==3.13.1
wheel==0.37.1
dlib==19.23.0
face-recognition-models==0.3.0
face-recognition==1.3.0
numpy==1.22.2
opencv-python==4.5.5.62
Pillow==9.0.1
pip==22.1.1
pytz==2022.1
setuptools==60.2.0
sqlparse==0.4.2
tzdata==2022.1
